Update and rename union.c to Exercicio 7.6/union.c

+++ Exercicio 7.6/union.c
@@ -1,3 +1,4 @@
+
 #include<stdio.h>
 #include<stdlib.h>
 
@@ -12,8 +13,8 @@ struct Portador{
     char nome[30];
     int  t;
     union {
-	char jogador[256];
-	float identificador;
+        char jogador[256];
+        double identificador;
     }id;
     struct Orb o;
 };
@@ -21,7 +22,7 @@ void preenche(struct Orb* orb){
     printf("Escolha o elmento da orb, o estilo e o seu nivel, respectivamente:\n");
     scanf("%s", (*orb).elem);
     scanf("%s", (*orb).estilo);
-    scanf("%d", (&(*orb).nivel));  
+    scanf("%d", (&(*orb).nivel));
 }
 
 void preenche2(struct Portador* p1){
@@ -31,11 +32,11 @@ void preenche2(struct Portador* p1){
     printf("Digite 1 para usar seu nome como identificador ou 2 para usar seu numero de sócio:\n");
     scanf("%d", &(*p1).t);
     if ((*p1).t==1){
-	printf("Digite seu nome:\n");
-	scanf("%s", (*p1).id.jogador);
+        printf("Digite seu nome:\n");
+        scanf("%s", (*p1).id.jogador);
     }else if ((*p1).t==2){
-	printf("Digite seu número de sócio;\n");
-	scanf("%f", &(*p1).id.identificador);
+        printf("Digite seu número de sócio;\n");
+        scanf("%lf", &(*p1).id.identificador);
     }
     preenche((&(*p1).o));
 }
@@ -45,7 +46,7 @@ int main(void){
     if (p1.t==1){
     	printf("\nSeu nome:%s\nSeu personagem:\nClasse:%s\nNome:%s\n\nSua Orb:\nElemento:%s\nEstilo:%s\nNível:%d\n",p1.id.jogador, p1.classe, p1.nome, p1.o.elem, p1.o.estilo, p1.o.nivel);
     }else if(p1.t==2){
-	printf("\nSeu numero de sócio:%f\nSeu personagem:\nClasse:%s\nNome:%s\n\nSua Orb:\nElemento:%s\nEstilo:%s\nNível:%d\n",p1.id.identificador, p1.classe, p1.nome, p1.o.elem, p1.o.estilo, p1.o.nivel);
-    }	
+        printf("\nSeu numero de sócio:%.0lf\nSeu personagem:\nClasse:%s\nNome:%s\n\nSua Orb:\nElemento:%s\nEstilo:%s\nNível:%d\n",p1.id.identificador, p1.classe, p1.nome, p1.o.elem, p1.o.estilo, p1.o.nivel);
+    }
     return 0;
 }
